+++ apps/cdn/CHANGELOG.md
@@ -1,5 +1,18 @@
 # cdn
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies and fixed a loading state issue on app
+- Updated dependencies
+    - @xernerx/components@0.2.21
+    - @xernerx/feedback@0.0.22
+    - @xernerx/lib@0.5.3
+    - @xernerx/providers@0.3.14
+    - @xernerx/styles@0.1.10
+    - @xernerx/ui@0.3.5
+
 ## 1.0.4
 
 ### Patch Changes